pyproject.toml
[tool.poetry]
name = "sjbillingclient"
version = "1.0.0"
description = "A Python wrapper for the Google Play Billing Library that facilitates in-app purchases and subscriptions in Android applications"
authors = ["Kenechukwu Akubue <kengoon19@gmail.com>"]
readme = "README.md"
homepage = "https://github.com/SimpleJnius/sj-android-billingclient"
repository = "https://github.com/SimpleJnius/sj-android-billingclient.git"
documentation = "https://github.com/SimpleJnius/sj-android-billingclient/blob/master/README.md"

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/SimpleJnius/sj-android-billingclient/issues"
"Source" = "https://github.com/SimpleJnius/sj-android-billingclient"
"Pull Requests" = "https://github.com/SimpleJnius/sj-android-billingclient/pulls"

[tool.poetry.dependencies]
python = "^3.9"
pyjnius = "^1.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
